[Bugfix:InstructorUI] Remove lowest null handling (#10680)
Currently, when web-based GUI customization page is rendered, users may
get below screenshot image warning (screenshot credit to @zacharymnp)
![스크린샷 2024-07-09 오전 11 07
16](https://github.com/Submitty/Submitty/assets/123261952/dbd09bca-c620-435f-b557-72ecf01caa38)


This is occurring because when we are loading json, it is looking for
remove_lowest even when there is none.

In order to reproduce/test this issue, go to reports page ->
customization page -> drag few gradeable to assigned bucket. Hit save
changes. Refresh page, and will see the screenshot warning.

With this change, it checks whether remove_lowest is specified or not,
and if it is not, it would set it to zero, and only parse when it is
specified.

---------

Co-authored-by: William Allen <16820599+williamjallen@users.noreply.github.com>