[Bugfix:TAGrading] Fix Grade Inquiry Toggle Button (#10701)
Closes #10576 
### What is the current behavior?
There is a lot of extra code, and the Grade Inquiry Only button for
grading is broken.

### What is the new behavior?
The grade inquiry function was being run when there was no display for
students yet, so I moved it to the bottom. Additionally, some of the
logic and code were extra and backwards so I changed them accordingly.