[Feature:Submission] Render Jupyter output images (#10751)
### What is the current behavior?
#10648 added bare-bones support for viewing Jupyter notebooks in
Submitty natively. As of that PR, only text output is supported.

### What is the new behavior?
This PR adds image output rendering to our Jupyter notebook support. As
per the
[docs](https://jupyterlab.readthedocs.io/en/1.2.x/user/file_formats.html),
we still need to implement several other types of output to be able to
fully support Jupyter notebooks. For now, the most common use cases have
been covered.


![image](https://github.com/user-attachments/assets/eedc03a6-65b9-42da-a6e4-7890f788617b)