[Bugfix:Submission] Improve daylight savings warning banner (#10721)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Closes #10682 and closes #10598. Currently the daylight savings banner
is displayed way too much, 1 week before daylight savings and one week
after daylight savings.

### What is the new behavior?
This PR addresses the display of the daylight savings banner, with it
using late days to calculate when they should show the banner. It also
aims to improve the current cypress checks for late days.

### Other information?
For the people reviewing:
There are a lot of conditions because of confusing late day numbers such
as 150 and 250, where it could escape daylight savings and then walk
back into daylight savings, making PHP say there is no daylight saving
that occured.
Because it is confusing to test 150 and 250, they are not included in
the cypress tests.

Cypress:
I tested 2024-03-09 (daylight savings is 3-10) at 0:00, 1:59, and 2:01
manually and it worked exactly as intended. 1:59 should not display a
banner for 1 late day because it is before the daylight change, same
with 0:00. 2:01 should trigger the banner, which it does.