[Bugfix:Developer] Fix abandoned tagger & remove warning (#10718)
### What is the current behavior?
Fixes #10609. Currently, the abandoned PR tagger checks if a PR has been
already given a warning and has had two days of inactivity. When an
abandoned PR is taken over, the tagger will re-apply the abandoned tag
after 2 days instead of 14 days of inactivity. This is undesirable.

### What is the new behavior?
This PR removes the warning message and simply has the script tag a PR
after 14 days of inactivity.