[Bugfix:Forum] Remove duplicated forum test (#10878)
### What is the current behavior?
A duplicate Cypress forum test was added, and this is unneeded. 

### What is the new behavior?
This test has been removed to reduce testing time, and possible
confusion later on on which file to add tests to.

test_forum.spec.js was the original Forum test created last summer, but
with all of the forum tests that have been created recently, they were
put in forums.spec.js, with the test from test_forum.spec.js copied, so
test_forum.spec.js is now obsolete.