[Bugfix:Submission] Fixing Alert to Explain Locked Gradeable (#10950)
Closes #10894 

### What is the current behavior?
When a gradeable is locked for a particular student, they do not know
what prior gradeable they need to complete before being able to access
the current one.

### What is the new behavior?
When a student hovers over the locked gradeable button, they see exactly
what prior gradeable they need to complete and the score they need to
receive on said gradeable.


![image](https://github.com/user-attachments/assets/e53699b6-0db2-4515-991e-7b2d138128d1)


### Other information?
To test, make a new gradeable with a prerequisite and required minimum
score as an instructor. Then, as a student, hover over or click on that
gradeable's "LOCKED" button.