[Bugfix:TAGrading] Wait for all Ajax calls to complete (#10984)
### What is the current behavior?
Currently on different browsers, it either waits or doesn't wait for all
the ajax calls to complete. This would result in a race condition where
the code was getting components that weren't saved, and would not save
it. From a graders standpoint, unless they had console open, they would
not get an error.

Another aspect of this has to do with the different ways the browser
deals with event loops, which chrome apparently just solves all async
requests and then moves to the next page. However, Firefox and Safari
casually just moves on while the async requests happen, which messes
with the code as it is also looking for elements on the screen. At least
this is my interpretation of the issue

partial(?) work for #10973 

### What is the new behavior?
Properly wait for all the ajax calls to complete. This is done through a
callback that only executes after it is done, so the event loop does not
matter anymore.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
Tested on Firefox, Chrome, and Safari