[Bugfix:Submission] Fixing Buggy Late Day Message Text (#10981)
### What is the current behavior?
Closes #10963 

The late day messages currently do not display a link to the "My Late
Days / Extensions" page and says "You have 1 remaining late day for this
assignment" instead of for future assignments.

### What is the new behavior?

The code fixes the above behavior:

![image](https://github.com/user-attachments/assets/07ddbb25-f836-461a-aa5c-30569c50975f)

In the image above, the instructor makes a gradeable that has a limit of
3 late days allowed. The student uses one late day, and the code now
displays that the student has two more late days remaining for that same
assignment. Also, it displays a link at the bottom to the 'My Late Days
/ Extensions' page.

### Other information?
To test, make or go to a late gradeable, and check if the above changes
have been made correctly.