[Refactor:Forum] Create doctrine forum entities (#10999)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->

### What is the new behavior?

This is part one of refactoring the forum rendering to use Doctrine.
This contains the entities and repositories that will be necessary to
render the forum with doctrine. It alone will not affect any forum
behavior. Part 2 will come after this and contain changes to the
controller, view, and frontend.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->